requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
PyPDF2==3.0.1
Pillow==10.0.1
python-magic==0.4.27
APScheduler==3.10.4
email-validator==2.1.0